small change for changing indices to u32

+++ crates/mani/fabled_render/src/mesh/container/mesh_data.rs
@@ -4,6 +4,6 @@ use crate::mesh::container::Vertex;
 #[repr(C, align(16))]
 pub struct Mesh {
     pub vertices: Vec<Vertex>,
-    pub indices: Vec<usize>,
+    pub indices: Vec<u32>,
     pub material_id: u32,
 }
